moved experimental features into a "playground" page
